+++ Model/Proyectil.cpp
@@ -9,8 +9,8 @@ Proyectil::Proyectil() {
     Proyectil::setPosicion(0,0);
 }
 
-Proyectil::Proyectil(Vector2d<float> unitario , Vector2d<int> posicion) {
-    this->velocidad = Vector2d<int>(unitario.getY() * rapi, unitario.getY() * rapi);
+Proyectil::Proyectil(Vector2d<float> unitario , Vector2d<float> posicion) {
+    this->velocidad = Vector2d<float>(unitario.getY() * rapi, unitario.getY() * rapi);
     this->posicion = posicion;
     Proyectil::setDimension(10,20);
     Proyectil::setPosicionCentral(this->posicion.getX()+5,this->posicion.getY()+10);
